fastapi study/gmic-web-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
tortoise-orm>=0.20.0
python-multipart>=0.0.6
openai>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
aiofiles>=23.2.1
python-magic>=0.4.27
requests>=2.31.0
asyncmy>=0.2.7